Complete refactoring of services into smaller module
